most popular movies api implementation

+++ android/app/src/main/java/com/backbase/assignment/ui/mostpopular/MostPopularMovie.kt
@@ -5,5 +5,5 @@ data class MostPopularMovie(
     val movieTitle: String,
     val movieOverView: String,
     val movieImageUrl: String,
-    val movieRating: Int
+    val movieRating: Double
 )
